+++ python/paddle/fluid/layers/utils.py
@@ -274,31 +274,3 @@ def _contain_var(list_or_tuple):
         if isinstance(item, Variable):
             return True
     return False
-
-
-def get_new_shape_tensor(list_shape):
-    new_shape_tensor = []
-    for dim in list_shape:
-        if isinstance(dim, Variable):
-            dim.stop_gradient = True
-            new_shape_tensor.append(dim)
-        else:
-            assert (isinstance(dim, int))
-            temp_out = helper.create_variable_for_type_inference('int64')
-            fill_constant([1], 'int64', dim, force_cpu=True, out=temp_out)
-            new_shape_tensor.append(temp_out)
-    return new_shape_tensor
-
-
-def get_attr_shape(list_shape):
-    unk_dim_idx = -1
-    attrs_shape = []
-    for dim_idx, dim_size in enumerate(list_shape):
-        if isinstance(dim_size, Variable):
-            attrs_shape.append(-1)
-        else:
-            attrs_shape.append(dim_size)
-            assert dim_size > 0, (
-                "Each dimension size given in shape must not be negative "
-                "except one unknown dimension.")
-    return attrs_shape
